update project structure.Fix db
