+++ MVP-project/backend/app/models/__init__.py
@@ -1,0 +1,2 @@
+from .userBase import User
+from .tokenBase import Token
