+++ MVP-project/backend/app/models/userBaseModel.py
@@ -1,0 +1,11 @@
+from pydantic import EmailStr, BaseModel
+
+class UserRegister(BaseModel):
+    name: str
+    password: str
+    email: EmailStr
+    role: str
+
+class UserLogin(BaseModel):
+    email: EmailStr
+    password: str
