+++ MVP-project/pyproject.toml
@@ -16,6 +16,7 @@ python-jose = "^3.3.0"
 uuid = "^1.30"
 sqlalchemy = "^2.0.36"
 psycopg2-binary = "^2.9.10"
+python-dotenv = "^1.0.1"
 
 
 [build-system]
